fix(front): logout
