+++ umanz_app/pages/index.vue
@@ -1,158 +1,62 @@
-<script setup lang="ts">
-definePageMeta({
-    layout: 'blank',
-    middleware: 'logout',
-    key: 'login-page'
-});
-
-import { z } from 'zod'
-import type { FormSubmitEvent } from '#ui/types'
-import { useAuth } from '~/features/auth/useAuth';
-
-const schema = z.object({
-    email: z.string().email('Email invalide'),
-    password: z.string().min(8, 'Le mot de passe doit contenir au moins 8 caractères')
-});
-type Schema = z.output<typeof schema>;
-
-const formState = reactive({
-    email: '',
-    password: '',
-    error: '',
-    loading: false, // in the future
-    remember: false // in the future
-});
-
-const isPasswordVisible = ref(false);
-const { login, setSession } = useAuth();
-
-async function onSubmit(event: FormSubmitEvent<Schema>) {
-    formState.error = '';
-    formState.loading = true;
-
-    try {
-        const apiUrl = useRuntimeConfig().public.apiUrl;
-        const loginResponse = await login(formState.email, formState.password, apiUrl);
-
-        if (loginResponse) {
-            setSession(loginResponse);
-            await navigateTo('/Home');
-        } else {
-            formState.error = 'Email ou mot de passe incorrect';
-        }
-    } catch (err: any) {
-        console.error('Login error:', err);
-        if (err.status === 404) {
-            formState.error = 'Email ou mot de passe incorrect';
-        } else {
-            formState.error = 'Une erreur est survenue lors de la connexion. Veuillez réessayer.';
-        }
-    } finally {
-        formState.loading = false;
-    }
+<script>
+export default {
+  name: 'Home',
+  // methods: {
+  //   register() {
+  //     this.$router.push( '/register' );
+  //   },
+  //   login() {
+  //     this.$router.push( '/login' );
+  //   }
+  // }
 }
 </script>
 
 <template>
-    <div :key="$route.fullPath" class="min-h-screen flex items-center justify-center bg-slate-50 relative overflow-hidden px-4">
-        <div class="w-full max-w-md transition-all duration-500 hover:translate-y-[-2px]">
-            <div class="bg-white/80 backdrop-blur-xl p-8 md:p-10 rounded-[2rem] shadow-2xl shadow-slate-200/50 border border-white/20">
-                <div class="text-center mb-8">
-                    <div class="inline-flex items-center justify-center w-16 h-16 rounded-[1.25rem] shadow-lg shadow-umanz-green/25 text-umanz-green text-3xl font-bold mb-6 transform transition-transform hover:scale-110 duration-300">
-                        U
-                    </div>
-                    <h2 class="text-3xl font-bold text-slate-900 tracking-tight">Connexion</h2>
-                    <p class="text-slate-500 mt-2 font-medium">Bon retour sur <span class="text-umanz-purple">UmanZ</span></p>
-                </div>
-
-                <UForm :schema="schema" :state="formState" class="space-y-5" @submit="onSubmit">
-                    <UFormGroup label="Email" name="email" :ui="{ label: { base: 'text-red-500 font-semibold mb-1.5' } }">
-                        <UInput 
-                            v-model="formState.email" 
-                            icon="i-heroicons-envelope" 
-                            placeholder="email@exemple.com"
-                            type="email"
-                            autocomplete="email"
-                            size="lg"
-                            variant="outline"
-                            :ui="{ rounded: 'rounded-2xl' }"
-                            class="transition-all duration-200"
-                        />
-                    </UFormGroup>
-
-                    <UFormGroup label="Mot de passe" name="password" :ui="{ label: { base: 'text-red-500 font-semibold mb-1.5' } }">
-                        <UInput 
-                            v-model="formState.password" 
-                            :type="isPasswordVisible ? 'text' : 'password'" 
-                            autocomplete="current-password"
-                            icon="i-heroicons-lock-closed" 
-                            placeholder="••••••••"
-                            size="lg"
-                            variant="outline"
-                            :ui="{ rounded: 'rounded-2xl', trailing: { padding: { lg: 'pe-12' } } }"
-                            class="transition-all duration-200"
-                        >
-                            <template #trailing>
-                                <UButton
-                                    color="gray"
-                                    variant="ghost"
-                                    :icon="isPasswordVisible ? 'i-heroicons-eye-slash' : 'i-heroicons-eye'"
-                                    class="mr-1"
-                                    size="sm"
-                                    @click="isPasswordVisible = !isPasswordVisible"
-                                />
-                            </template>
-                        </UInput>
-                    </UFormGroup>
-
-                    <div class="flex items-center justify-between">
-                        <UCheckbox 
-                            v-model="formState.remember" 
-                            label="Se souvenir de moi"
-                            :ui="{ label: 'text-sm font-medium text-slate-600', base: 'h-4 w-4 rounded-md border-slate-300 text-umanz-purple focus:ring-umanz-purple' }"
-                        />
-                        <ULink 
-                            to="/forgot-password" 
-                            class="text-sm font-semibold text-umanz-purple hover:text-umanz-purple/80 transition-colors"
-                        >
-                            Mot de passe oublié ?
-                        </ULink>
-                    </div>
-
-                    <div>
-                        <p v-if="formState.error" class="text-red-600 text-sm font-medium text-center bg-red-50 py-3 px-4 rounded-2xl border border-red-100 mb-2">
-                            {{ formState.error }}
-                        </p>
-                    </div>
-
-                    <UButton 
-                        type="submit" 
-                        block 
-                        size="xl" 
-                        color="primary"
-                        :loading="formState.loading"
-                        :disabled="formState.loading"
-                        class="py-4 font-bold text-base transition-all duration-300"
-                        :ui="{ rounded: 'rounded-2xl' }"
-                    >
-                        Me connecter
-                    </UButton>
-
-                    <div class="text-center text-sm text-slate-500 mt-8 font-medium">
-                        Vous n'avez pas encore de compte ?
-                        <ULink 
-                            class="text-umanz-green font-bold hover:text-umanz-green/80 transition-colors ml-1" 
-                            to="/register"
-                        >
-                            Inscrivez-vous ici
-                        </ULink>
-                    </div>
-                </UForm>
-            </div>
-            
-            <p class="text-center text-slate-400 text-xs mt-8">
-                &copy; {{ new Date().getFullYear() }} UmanZ. Tous droits réservés.
-            </p>
-        </div>
+  <div class="space-y-8">
+    <div class="text-center">
+      <h1 class="text-4xl font-extrabold text-gray-900 mb-4 tracking-tight">Bienvenue sur <span class="text-umanz-purple">UmanZ</span></h1>
+      <p class="text-gray-500 text-lg max-w-2xl mx-auto">
+        Gérez vos ressources humaines facilement et efficacement avec notre plateforme unifiée.
+      </p>
     </div>
+
+    <div class="grid grid-cols-1 md:grid-cols-3 gap-6">
+      <UCard class="hover:border-umanz-purple transition-colors cursor-pointer group">
+        <template #header>
+          <div class="flex items-center gap-3">
+            <div class="w-10 h-10 rounded-xl bg-umanz-purple/10 flex items-center justify-center text-umanz-purple">
+              <UIcon name="i-heroicons-user-group" class="text-2xl" />
+            </div>
+            <h3 class="font-bold text-gray-900">Employés</h3>
+          </div>
+        </template>
+        <p class="text-sm text-gray-500">Gérez vos talents et leurs informations personnelles.</p>
+      </UCard>
+
+      <UCard class="hover:border-umanz-green transition-colors cursor-pointer group">
+        <template #header>
+          <div class="flex items-center gap-3">
+            <div class="w-10 h-10 rounded-xl bg-umanz-green/10 flex items-center justify-center text-umanz-green">
+              <UIcon name="i-heroicons-calendar" class="text-2xl" />
+            </div>
+            <h3 class="font-bold text-gray-900">Congés</h3>
+          </div>
+        </template>
+        <p class="text-sm text-gray-500">Suivez les demandes de congés et les absences.</p>
+      </UCard>
+
+      <UCard class="hover:border-umanz-orange transition-colors cursor-pointer group">
+        <template #header>
+          <div class="flex items-center gap-3">
+            <div class="w-10 h-10 rounded-xl bg-umanz-orange/10 flex items-center justify-center text-umanz-orange">
+              <UIcon name="i-heroicons-megaphone" class="text-2xl" />
+            </div>
+            <h3 class="font-bold text-gray-900">Annonces</h3>
+          </div>
+        </template>
+        <p class="text-sm text-gray-500">Diffusez des offres et gérez les recrutements.</p>
+      </UCard>
+    </div>
+  </div>
 </template>
